Add basic test framework for instruction assembly
The tests are generated at runtime based on a template
(assemblerCases), but always the same. This is because every assembling
test case has a corresponding, nearly identical dissasembling test case.

I used copy/paste to create the ADD and AND tests, but that may change
in the future if I decide to add simple generators like with the toBits.